pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "osint-mcp"
version = "0.1.0"
description = "Continuous OSINT / attack-surface monitoring for bug bounty hunting, exposed as an MCP server with Discord notifications."
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "bleucpu" }]
license = { text = "MIT" }
dependencies = [
    "mcp>=1.2.0",
    "httpx>=0.27.0",
    "feedparser>=6.0.11",
    "pyyaml>=6.0.1",
    "python-dateutil>=2.9.0",
    "apscheduler>=3.10.4",
    "aiosqlite>=0.20.0",
    "beautifulsoup4>=4.12.3",
    "lxml>=5.2.0",
    "python-dotenv>=1.0.0",
    "websockets>=12.0",
]

[project.optional-dependencies]
bbot = ["bbot>=2.1.0"]

[project.scripts]
osint-mcp = "osint_mcp.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
osint_mcp = ["templates/*.yaml"]
